refactor: relocate ColorScheme component and update import paths
- Moved ColorScheme.vue from the widget-builder directory to a new color-scheme subdirectory for better organization.
- Updated import statements in WidgetLiveChat.vue to reflect the new path.
- Adjusted the template structure in WidgetLiveChat.vue for improved readability.
